backend/requirements.txt
flask==2.3.2
flask-cors
google-api-python-client==2.85.0
google-auth==2.18.0
google-cloud-logging==3.5.0
google-cloud-bigquery==3.10.0
google-auth-oauthlib 
google-auth-httplib2 
google-cloud-storage
pandas